feat: split group and aggregation transformers
